Fix bug where a provisioner disconnect would cause collectors to purge mediators
This wouldn't happen the first time a provisioner disconnected,
but any subsequent disconnections would result in an unexpected
purge.

The problem was caused by the "flagtimer" which had been set from
the first reconnection, but does not get reset until the second
reconnection. However, because our code path checks if the timer
has expired before the provisioner can reconnect, we were
erroneously assuming that the timer had immediately fired and
started purging any unconfirmed mediators (i.e. all mediators,
as they had just been flagged as unconfirmed when the provisioner
went away).

Now, we make a point to halt the flagtimer as soon as we see the
provisioner disconnect, and also take care not to trigger any
actions even if that halted timer does have an epoll event.

+++ src/collector/collector_forwarder.c
@@ -287,6 +287,11 @@ static void flag_all_destinations(forwarding_thread_data_t *fwd) {
         med->awaitingconfirm = 1;
     }
     fwd->awaitingconfirm = 1;
+
+    if (fwd->flagtimerfd != -1) {
+        close(fwd->flagtimerfd);
+        fwd->flagtimerfd = -1;
+    }
 }
 
 static int handle_ctrl_message(forwarding_thread_data_t *fwd,
@@ -701,7 +706,7 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
     /* Add the mediator confirmation timer to our poll item list, if
      * required.
      */
-    if (fwd->awaitingconfirm) {
+    if (fwd->awaitingconfirm && fwd->flagtimerfd != -1) {
         fwd->topoll[fwd->nextpoll].socket = NULL;
         fwd->topoll[fwd->nextpoll].fd = fwd->flagtimerfd;
         fwd->topoll[fwd->nextpoll].events = ZMQ_POLLIN;
@@ -751,7 +756,7 @@ static inline int forwarder_main_loop(forwarding_thread_data_t *fwd) {
         fwd->topoll[1].revents = 0;
     }
 
-    if (fwd->awaitingconfirm) {
+    if (fwd->awaitingconfirm && fwd->flagtimerfd != -1) {
         if (fwd->topoll[fwd->nextpoll].revents & ZMQ_POLLIN) {
             purge_unconfirmed_mediators(fwd);
             fwd->awaitingconfirm = 0;
